Added framework for running experiments.
Experiment class is a manager for individual runs of an experiment.
Options are passed via a JSON config file.

Scene class is an individual run which manages meshes and reacts to input.
Logic goes here.

Reworked the MouseKeyboardInterface class to better fit with paradigm used
by Experiment.  Other input types may need to run in their own thread.

TODO:
- add logging to Scene.
- more input types
- should we have a 'replay' input type?

+++ Experiment/D3DRenderer.cpp
@@ -341,9 +341,9 @@ void D3DRenderer::DrawWindowTarget(D3DWindowTarget* wt, int x, int y) {
 	if (x != wt->x || y != wt->y) {
 		wt->x = x;
 		wt->y = y;
-		float left = x;
+		float left = (float)x;
 		float right = left + (float)wt->w;
-		float top = y;
+		float top = (float)y;
 		float bottom = top + (float)wt->h;
 		D3DPoint vertices[nwindowverts];
 		// First triangle.
